upload and download files

+++ controllers/admin.js
@@ -4,6 +4,8 @@ const { validationResult } = require("express-validator");
 
 const Product = require("../models/product");
 
+const fileHelper = require("../utils/file");
+
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
     pageTitle: "Add Product",
@@ -142,6 +144,10 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       product.imageUrl = updatedImageUrl && updatedImageUrl.path;
+
+      if (updatedImageUrl) {
+        fileHelper.deleteFile(product.imageUrl);
+      }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
@@ -175,14 +181,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        next(new Error("Product not found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      const error = new Error(err);
-      err.httpStatusCode = 500;
-      return next(error);
+      next(err);
     });
 };
